+++ adapters/Facebook/FacebookAdapter/OpenBidding/GADFBRewardedInterstitialRenderer.h
@@ -1,4 +1,4 @@
-// Copyright 2016 Google Inc.
+// Copyright 2020 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,20 +13,10 @@
 // limitations under the License.
 
 #import <Foundation/Foundation.h>
+#import <GoogleMobileAds/GoogleMobileAds.h>
 
-#import "GADMChartboostExtras.h"
+#import "GADFBRewardedRenderer.h"
 
-/// The purpose of the GADMAdapterChartboostDataProvider protocol is to allow the singleton to
-/// interact with the adapter.
-@protocol GADMAdapterChartboostDataProvider <NSObject>
-
-/// Returns the Chartboost extras object.
-- (GADMChartboostExtras *)extras;
-
-/// Returns the Chartboost ad location.
-- (NSString *)getAdLocation;
-
-/// Called after an ad has failed to load.
-- (void)didFailToLoadAdWithError:(NSError *)error;
+@interface GADFBRewardedInterstitialRenderer : GADFBRewardedRenderer
 
 @end