+++ adapters/Tapjoy/TapjoyAdapter/GADMediationAdapterTapjoy.m
@@ -44,7 +44,7 @@
 
   if (!sdkKeys.count) {
     NSError *error = GADMAdapterTapjoyErrorWithCodeAndDescription(
-        kGADErrorMediationDataError,
+        GADMAdapterTapjoyErrorInvalidServerParameters,
         @"Tapjoy mediation configurations did not contain a valid SDK Key.");
     completionHandler(error);
     return;